refactor: extract artifact bundle verifier
